Refactoring Header
