+++ front-end/app/layout/header/headerComponent.module.css
@@ -99,212 +99,6 @@ header#background button svg {
     }
 }
 
-/* SEARCH BUTTON SECTION ---> DESKTOP and MOBILE */
-#search_container #form_search form input,
-#search_container #form_mobile_search form input {
-    padding: 12px 12px;
-    padding-left: 18px;
-
-    border-radius: 24px 0 0 24px;
-    border: 1px solid var(--black-75);
-    border-right: 0;
-
-    background-color: var(--white-100);
-}
-
-#search_container #form_search form input:focus,
-#search_container #form_mobile_search form input:focus {
-
-    /* outline: solid; */
-    outline-width: 0;
-    /* outline-color: var(--brand-color); */
-
-    border: 1px solid var(--brand-color);
-    border-right: 0;
-}
-
-#search_container #form_search form button,
-#search_container #form_mobile_search form button {
-
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    padding: 8px 12px;
-    border-radius: 0 24px 24px 0;
-
-    border: 1px solid var(--black-75);
-    border-left: 0;
-
-    background-color: var(--white-100);
-
-}
-
-#search_container #form_search form button:hover svg,
-#search_container #form_mobile_search form button:hover svg {
-
-    fill: var(--brand-color);
-
-}
-
-#search_container #form_search form input:focus~button,
-#search_container #form_mobile_search form input:focus~button {
-    border: 1px solid var(--brand-color);
-    border-left: 0;
-}
-
-#search_container #form_search form button svg,
-#search_container #form_mobile_search form button svg {
-
-    transform: scale(1.21);
-
-    fill: var(--black-50);
-
-    margin-bottom: 4px;
-
-}
-
-#form_mobile_search {
-
-    position: absolute;
-    top: 60px;
-    bottom: 0;
-    left: 0;
-    right: 0;
-
-    background-color: var(--background);
-
-}
-
-#form_mobile_search>form {
-
-    width: max-content;
-
-    margin: 0 auto;
-
-}
-
-#search_container>#form_mobile_search[aria-expanded=true] {
-
-    animation: openForm forwards 200ms;
-
-}
-
-#search_container>#form_mobile_search[aria-expanded=false] {
-
-    animation: closeForm forwards 200ms;
-
-}
-
-/* SEARCH FORM*/
-@keyframes openForm {
-    0% {
-        height: 0;
-        opacity: 0;
-    }
-
-    100% {
-        opacity: 1;
-        z-index: 10;
-        height: 60px;
-    }
-}
-
-@keyframes closeForm {
-    0% {
-        height: 60px;
-        opacity: 1;
-    }
-
-    100% {
-        opacity: 0;
-        z-index: 0;
-        display: none;
-        height: 0;
-    }
-}
-
-/* SEARCH RESULTS CONTAINER*/
-#search_results_container {
-
-    z-index: 2;
-
-    position: absolute;
-    top: 120px;
-    left: 0;
-    right: 0;
-
-    background: linear-gradient(var(--background) 70%, var(--background));
-}
-
-@media(min-width: 550px) {
-    #search_results_container {
-        top: 67px;
-    }
-}
-
-@media(min-width: 1020px) {
-    #search_results_container {
-        left: auto;
-        width: 50vw;
-    }
-}
-
-@media(min-width: 1440px) {
-    #search_results_container {
-        width: 40vw;
-    }
-}
-
-#search_results_container>button {
-
-    display: block;
-    margin-top: 16px;
-    margin-left: auto;
-    margin-right: auto;
-    margin-bottom: 24px;
-
-    text-align: center;
-
-    border: none;
-
-    border-bottom: 2px solid var(--white-75);
-    background: transparent;
-
-    color: var(--white-75);
-
-    font-weight: 600;
-
-}
-
-#search_results_container>ul {
-
-    padding: 16px 0;
-
-    width: 96%;
-    margin: auto;
-    display: flex;
-    flex-direction: column;
-    gap: 4px 0;
-
-}
-
-@media(min-width: 550px) {
-    #search_container button#btn_open_search_form_mobile {
-
-        display: none;
-
-    }
-}
-
-@media(max-width: 549px) {
-    #search_container #form_search {
-
-        display: none;
-
-    }
-}
-
 #navbar_container {
 
     width: 100%;
